oris/backend/djangoDoda/appProductMetrics/requirements.txt
django
djangorestframework
django-cors-headers
elasticsearch
pandas
requests
transformers
torch